agents/materials-agent/requirements.txt
fastapi
uvicorn
openai
google-generativeai
pydantic
python-dotenv
requests
